removed unnecessary files, pep8 fixes
