--- eventkit_cloud/api/templates/rest_framework/api.html
@@ -1,29 +1,0 @@
-{% extends "rest_framework/base.html" %}
-{% load staticfiles %}
-{% block head %}
-
-            {% block meta %}
-                <meta http-equiv="Content-Type" content="text/html; charset=utf-8"/>
-                <meta name="robots" content="NONE,NOARCHIVE" />
-            {% endblock %}
-
-            <title>{% block title %}HOT Export API Docs{% endblock %}</title>
-
-            {% block style %}
-                {% block bootstrap_theme %}
-                    <link rel="stylesheet" type="text/css" href="{% static "rest_framework/css/bootstrap.min.css" %}"/>
-                    <link rel="stylesheet" type="text/css" href="{% static "rest_framework/css/bootstrap-tweaks.css" %}"/>
-                {% endblock %}
-                <link rel="stylesheet" type="text/css" href="{% static "rest_framework/css/prettify.css" %}"/>
-                <link rel="stylesheet" type="text/css" href="{% static "rest_framework/css/default.css" %}"/>
-                <link rel="stylesheet" type="text/css" href="{% static "css/style.css" %}"/>
-            {% endblock %}
-
-{% endblock %}
-
-{% block branding %}
-    <a class='navbar-brand' rel="nofollow" href="">
-        HOT Export API Documentation 2.0.0
-    </a>
-{% endblock %}
-
