--- eventkit_cloud/api/templates/rest_framework/login.html
@@ -1,3 +1,0 @@
-{% extends "rest_framework/login_base.html" %}
-
-{# Override this template in your own templates directory to customize #}
