+++ eventkit_cloud/settings/base.py
@@ -15,9 +15,6 @@ from .secret import SECRET_KEY  # NOQA  # isort:skip
 # In a Windows environment this must be set to your system time zone.
 TIME_ZONE = 'UTC'
 
-# default DEBUG setting
-DEBUG = True
-
 # Language code for this installation. All choices can be found here:
 # http://www.i18nguy.com/unicode/language-identifiers.html
 LANGUAGE_CODE = 'en'
